Forms: Validation errors

+++ TreeCraftingVoyager.Server/Controllers/CategoriesController.cs
@@ -1,20 +1,13 @@
 ﻿using AutoMapper;
-using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Newtonsoft.Json;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http.Json;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using TreeCraftingVoyager.Server.Data.Repositories.Crud;
 using TreeCraftingVoyager.Server.Data.Repositories.Tree;
 using TreeCraftingVoyager.Server.Models.Dto.Category;
 using TreeCraftingVoyager.Server.Models.Entities;
-using TreeCraftingVoyager.Server.Models.ViewModels;
-using TreeCraftingVoyager.Server.Services;
 
 namespace TreeCraftingVoyager.Server.Controllers
 {
@@ -23,26 +16,21 @@ namespace TreeCraftingVoyager.Server.Controllers
     public class CategoriesController : ControllerBase
     {
         private readonly IMapper _mapper;
-        private readonly ICategoryService _categoryService;
         private readonly ITreeRepository<Category, CategoryDto, UpdateCategoryDto, CreateCategoryDto> _treeRepository;
         private readonly ICrudRepository<Category, CategoryDto, UpdateCategoryDto, CreateCategoryDto> _crudRepository;
 
         public CategoriesController(
             IMapper mapper,
-            ICategoryService categoryService,
             ITreeRepository<Category, CategoryDto, UpdateCategoryDto, CreateCategoryDto> treeRepository,
-            ICrudRepository<Category, CategoryDto, UpdateCategoryDto, CreateCategoryDto> crudRepository
-            )
+            ICrudRepository<Category, CategoryDto, UpdateCategoryDto, CreateCategoryDto> crudRepository)
         {
             _mapper = mapper;
-            _categoryService = categoryService;
             _treeRepository = treeRepository;
             _crudRepository = crudRepository;
         }
 
-
         [HttpGet("GetCategories")]
-        public async Task<IActionResult> GetRootCategories()
+        public async Task<ActionResult<IEnumerable<CategoryDto>>> GetRootCategories()
         {
             var ret = await _treeRepository.GetRootObjects();
 
@@ -50,25 +38,27 @@ namespace TreeCraftingVoyager.Server.Controllers
         }
 
         [HttpGet("GetChildrens/{id}")]
-        public async Task<IActionResult> GetCategoryChildrens(long id)
+        public async Task<ActionResult<IEnumerable<CategoryDto>>> GetCategoryChildrens(long id)
         {
-            var ret = await _crudRepository.GetQuery(q => q.Where(e => e.Id == id).Include(e => e.Childrens)).SingleAsync();
+            var ret = await _crudRepository.GetQuery(q => q.Where(e => e.Id == id).Include(e => e.Childrens)).SingleOrDefaultAsync();
+            if (ret == null)
+                return NotFound();
 
-            return Ok(ret.Childrens);
+            return Ok(ret);
         }
 
         [HttpGet("GetHierarchy")]
-        public async Task<IActionResult> GetCategoriesHierarchy()
+        public async Task<ActionResult<IEnumerable<CategoryDto>>> GetCategoriesHierarchy()
         {
             var ret = await _treeRepository.GetAllRecursively();
-            
-            return Ok(ret.Where(e => e.ParentId == null));
+
+            return Ok(ret);
         }
 
         [HttpGet("Get")]
-        public async Task<IActionResult> GetCategories()
+        public async Task<ActionResult<IEnumerable<CategoryDto>>> GetCategories()
         {
-            var ret = await _categoryService.GetAllCategoriesIncludingProducts(); // po co IncludingProducts???
+            var ret = await _crudRepository.GetAllAsync();
 
             return Ok(ret);
         }
@@ -77,28 +67,34 @@ namespace TreeCraftingVoyager.Server.Controllers
         public async Task<ActionResult<CategoryDto>> GetCategory(long id)
         {
             var ret = await _crudRepository.GetByIdAsync(id);
+            if (ret == null)
+                return NotFound();
 
             return Ok(ret);
         }
 
         [HttpPost("Create")]
-        public async Task<ActionResult<CategoryDto>> CreateCategory([FromBody] CreateCategoryDto category)
+        public async Task<ActionResult> CreateCategory([FromBody] CreateCategoryDto createDto)
         {
-            if (ModelState.IsValid)
-                await _crudRepository.CreateAsync(category);
-            else
-                return ValidationProblem(ModelState);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
 
-            return Ok();
+            var result = await _crudRepository.CreateAsync(createDto);
+            if (result == null)
+                return StatusCode(500, "Błąd serwera podczas tworzenia kategorii");
+
+            return NoContent();
         }
 
         [HttpPut("Update")]
-        public async Task<IActionResult> UpdateCategory(UpdateCategoryDto category)
+        public async Task<IActionResult> UpdateCategory([FromBody] UpdateCategoryDto updateDto)
         {
-            if (ModelState.IsValid)
-                await _crudRepository.UpdateAsync(category);
-            else
-                return ValidationProblem(ModelState);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var result = await _crudRepository.UpdateAsync(updateDto);
+            if (result == null)
+                return NotFound("Kategoria nie została znaleziona");
 
             return Ok();
         }
@@ -106,9 +102,13 @@ namespace TreeCraftingVoyager.Server.Controllers
         [HttpDelete("Delete/{id}")]
         public async Task<IActionResult> DeleteCategory(long id)
         {
+            var result = await _crudRepository.GetByIdAsync(id);
+            if (result == null)
+                return NotFound("Kategoria nie została znaleziona");
+
             await _crudRepository.DeleteAsync(id);
 
-            return Ok();
+            return NoContent();
         }
     }
 }
